+++ src/main/java/org/folio/search/model/entity/InstanceSubjectEntityAgg.java
@@ -7,5 +7,7 @@ public record InstanceSubjectEntityAgg(
   String id,
   String value,
   String authorityId,
+  String sourceId,
+  String typeId,
   Set<InstanceSubResource> instances
 ) { }