+++ src/main/java/org/folio/search/model/index/SubjectResource.java
@@ -6,4 +6,6 @@ public record SubjectResource(
   String id,
   String value,
   String authorityId,
+  String sourceId,
+  String typeId,
   Set<InstanceSubResource> instances) { }