+++ src/main/java/org/folio/search/service/browse/ContributorBrowseService.java
@@ -2,6 +2,7 @@ package org.folio.search.service.browse;
 
 import static java.util.Objects.nonNull;
 import static org.folio.search.utils.SearchUtils.AUTHORITY_ID_FIELD;
+import static org.folio.search.utils.SearchUtils.MISSING_LAST_PROP;
 import static org.opensearch.index.query.QueryBuilders.boolQuery;
 import static org.opensearch.index.query.QueryBuilders.matchAllQuery;
 import static org.opensearch.index.query.QueryBuilders.termQuery;
@@ -32,7 +33,6 @@ import org.springframework.stereotype.Service;
 public class ContributorBrowseService extends
   AbstractBrowseServiceBySearchAfter<InstanceContributorBrowseItem, ContributorResource> {
 
-  private static final String MISSING_LAST_PROP = "_last";
   private static final String CONTRIBUTOR_NAME_TYPE_ID_FIELD = "contributorNameTypeId";
   private static final String CONTRIBUTOR_TYPE_ID_FIELD = "instances.typeId";